**BREAKING**: Merge the hostApp model into the OS model
Change-type: major
